+++ GRAPHQL_MULTIPART_REQUEST_SPEC_URL.mjs
@@ -1,7 +1,5 @@
 // @ts-check
 
-"use strict";
-
 /**
  * [GraphQL multipart request spec](https://github.com/jaydenseric/graphql-multipart-request-spec)
  * URL. Useful for error messages, etc.
@@ -9,4 +7,4 @@
 const GRAPHQL_MULTIPART_REQUEST_SPEC_URL =
   "https://github.com/jaydenseric/graphql-multipart-request-spec";
 
-module.exports = GRAPHQL_MULTIPART_REQUEST_SPEC_URL;
+export default GRAPHQL_MULTIPART_REQUEST_SPEC_URL;