+++ graphqlUploadKoa.test.mjs
@@ -5,8 +5,8 @@ import { createServer } from "http";
 import Koa from "koa";
 import fetch, { File, FormData } from "node-fetch";
 
-import graphqlUploadKoa from "./graphqlUploadKoa.js";
-import processRequest from "./processRequest.js";
+import graphqlUploadKoa from "./graphqlUploadKoa.mjs";
+import processRequest from "./processRequest.mjs";
 import listen from "./test/listen.mjs";
 
 /**
@@ -40,7 +40,7 @@ export default (tests) => {
     /**
      * @type {{
      *   variables: {
-     *     file: import("./Upload.js"),
+     *     file: import("./Upload.mjs").default,
      *   },
      * } | undefined}
      */
@@ -80,7 +80,7 @@ export default (tests) => {
       /**
        * @type {{
        *   variables: {
-       *     file: import("./Upload.js"),
+       *     file: import("./Upload.mjs").default,
        *   },
        * } | undefined}
        */